requirement.txt
python-dotenv
google-generativeai
requests
pandas
numpy